backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
pydantic-settings
sqlalchemy
aiosqlite
python-dotenv
openai
tenacity
greenlet
